Append the SNAT rule in management chain
On startup we are flushing all the rules in ovnk
chains. The MGTM chain particularly gets flushed
from `createPlatformManagementPort` which calls
`tearDownManagementPortConfig` before `setupManagementPortConfig`
that readds these rules back. This PR doesn't
fix the flushing bug, but fixes the ovn-k8s-mp0 SNAT
rule to be always added last in the chain (append)
on startup. The other rules in the chain are inserted
or prepend-ed so we should be good.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>