master: don't list node pods on add/update unless necessary
Instead of always listing all pods on a node even when we don't use
the result, just list them if we are going to iterate the pods to
ensure they are set up.

This fixes a performance regression where ovnkube-master puts more
pressure on the apiserver/etcd when node update events occur.

Signed-off-by: Dan Williams <dcbw@redhat.com>