ACL logging: rename and rewrite oc.aclLoggingCanEnable
Rename oc.aclLoggingCanEnable to aclLoggingUpdateNsInfo and make some
changes to this method and how it is being called.

When a namespace's ACL logging annotation is updated to either {}
or {"allow": "", "deny": ""}, make sure to disable ACL logging
completely.

When an invalid value is provided for the entire annotation or for
"allow" or "deny", reset logging to "" for either key or for both
and log an error.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>