Implemnet syncMap object with lockable keys.
Isolate syncMap implementation from the retry logic:
  - retryObjEntry doesn't need to have mutex anymore
  - retryObjs doesn't work with cache mutexes, only calls methods

syncMap is a map with lockable keys. It allows to lock the key
regardless of whether the entry for given key exists.
When key is locked other threads can't read/write the entry with
the key.

Split ensureRetryEntryLocked into loadOrStore that will
make sure newEntry is present and locked, and load that was
previously called as ensureRetryEntryLocked with nil newRetryEntry.
Change iterateRetryResources to only lock every entry once and not use
cache mutex. Create a snapshot of keys that will be retried instead
of holding map lock.

Lock retryObject key until work for this key is completed.
That also removes the need for .ignore field - retry loop
will wait until key is unlocked.

Signed-off-by: Nadia Pinaeva <npinaeva@redhat.com>