Add OVN controller southbound database connection metric
Report OVN controller connection status via a metric from each
OVN kube node.

The metric is updated every 2 minutes in-order not to incur
polling overhead which is not insignificant.

Prometheus scrape interval is usually every 30 seconds, and is
this value on OCP.

The reason we report this from every node instead of grabbing it
from SB DB via (nb_cfg trigger), is to save SB DB from each
OVN controller having to write to it.
This maybe an issue when scaling node.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>