cni: don't settle addresses or set IPv6 parameters for IPv4
If the pod only has IPv4 addresses, we don't need to do any IPv6
stuff because it won't matter anyway. The sysctls take the RTNL
lock which under heavy load contends with other users (metrics,
cadvisor, node exporter), and SettleAddresses() does netlink
operations every 50ms.

Even with all that, this doesn't have a huge affect in heavy
load cases because the OVS operations and waiting for ovn-installed
take way more time, and waiting for addresses to settle actually
runs in parallel with ovn-installed.

But let's not make the netlink problem worse, and let's credit
the time taken for operations to the rigth place instead of
masking it.

Signed-off-by: Dan Williams <dcbw@redhat.com>