EgressIP: Export and small refactor no-reroute policies
The no-reroute (allow) policies that ensure east<->west traffic
does not break when steering a pod's traffic needs to be
created the same for any controller that creates reroute policies
(like the egress service controller).
This change allows us to reuse the functions that create them
in these controllers.
Also, we add an ExternalID for the node policy so that a level driven
controller can pass the node's key to the delete path as it can't send
a deleted node's object.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>