ovn node, gw router: set node masquerade IP on br-ex
Since the `br-ex` bridge already has the node masquerade IP address, it
is no longer required to add the masquerade route to the nodes.

The SNAT/unSNAT traffic to/from the node to use the node masquerade IP
address will **still** be used, to specifically support the use case of
packets sent from the node having a different source IP address.

The `V4MasqueradeSubnet` had to be updated, since follow-up commits will
require the `.4` address on that subnet to use as a dummy next-hop for
deployments where there isn't a next-hop available.

netlink utils: rename  method

This `LinkRoutesAddOrUpdateMTU` function can now be used to update the
route MTU *or* the route source. Thus, rename the method accordingly.

A follow up commit should try to come up with a better abstraction,
since this function already does too much:
It either:
- adds a route **or**
- updates a route
  - MTU **or**
  - source

I think we should have 3 different functions, each doing a single thing,
and leave to the caller to know which of those to invoke.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
Co-authored-by: Tim Rozet <trozet@redhat.com>

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>