E2E: fix tests stuck waiting on a channel
These tests run checkContinuousConnectivity in a goroutine and
wait on the podChan to continue. The problem is that if the
goroutine fails to create a pod it sends a msg to the errChan
and exits, making the test hang indefinitely until the global
timeout kicks in (3hrs).
Fixing checkContinuousConnectivity to send a nil to the errChan
if the pod was created successfully, and the callers to wait
for that nil before trying to get the pod.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>