Fix pointer issues in gateway code
There are two places in the gateway code where
the pointer return values were not being done
correctly. Since we watch for endpoint slice
changes, each update of such an endpoint will
need to retrigger an update of iptable rules
for those services. This update is done as
a delete of old rules followed by re-adding
the new rules that reflect the change.

Since the old obj will not always be available
like during a service update we maintain a cache
that is used to specifically delete the old set
of rules for the service state. This is where
the pointer bug was because we weren't really returning
the existing state of cache but instead returning
the updated new state which was being used for
deletion.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>