Don't assume policy type when evaluating isolation
PolicyTypes field of a policy should always be set to an appropriate
value before it reaches the network plugin. The network plugin should
not make any assumptions about what the policy type is outside of what
PolicyTypes indicates towards evaluating the isolation status of a
selected pod.

Not doing any re-synchronization of possibly affected policies because:
- If the user does not provide PolicyTypes, api server will fill those
  appropriately.
- It is unlikely that a user provides PolicyTypes that do not include a
  type for which rules are provided as well. And if it does, it's
  probably better not to change the behavior of those policies anyway.

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>