Prepare KIND cluster for Topology Aware Hints
This PR adds topology zone annotation to all
the worker role nodes in the cluster. This triggers
the sig-network e2e for topology hints feature.
NOTE: Since HA modes have 0 worker nodes, this e2e
runs only in non-HA modes, where we have 2 worker
nodes.

NOTE: There are some specific set of conditions to run
topology aware hints. When we run e2e's we
consider control plane node to be a worker since
we don't taint the node with unscheduable taint.
This casuses zones:endpoints to be out of sync.
We need to run at least one lane with proper
control-plane and worker node distinction.

This PR adds the KEEP TAINT option to the non-HA
mode so that TAH e2e's can be run properly.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>