Implement Topology Aware Hints
TAH is a feature specifically related to "service" traffic.
It allows users to route traffic "originating from a specific
zone" to head towards the endpoints in the "same zone". If
there aren't any endpoints in that zone, unlike ITP or ETP
this feature will allow for traffic to go to all endpoints
in the cluster. So simply put this feature allows to keep
traffic within the zone where it originated from.

What is a "zone" here?
k8s node object has a well-known label called
topology.kubernetes.io/zone: "us-east-1c". See
https://kubernetes.io/docs/reference/labels-annotations-taints/#topologykubernetesiozone
for more information. This way all nodes in the cluster
can be seggregated into zones.

How to enable TAH on a service?
This can be done via setting an annotation on the service.
service.kubernetes.io/topology-aware-hints: "auto". See
https://kubernetes.io/docs/concepts/services-networking/topology-aware-hints/#using-topology-aware-hints
for more information.

Conditions that OVNK will look for to enable TAH feature are
the same as what kube-proxy does:
// * The "service.kubernetes.io/topology-aware-hints" annotation on this Service is set to "Auto"
// * The node's labels include "topology.kubernetes.io/zone"
// * All of the endpoints for this Service have a topology hint
// * At least one endpoint for this Service is hinted for this node's zone.

In OVNK, LBs are of two types: cluster wide and per node, as
soon as the above conditions kick in, we will filter out
endpoints belonging to the zone that the node lies in, thus
making TAH service LBs, the per node type.

NOTE: No docs are being added for this feature since
upstream k8s docs reflect all the info on this feature.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>