Add logging verbosity to configuring OVN logs
When updating a large number of address_set, logs get flooded with
"Configuring OVN" which cannot be reduced.
Also klog shows up in the profiling data as a source of CPU consumption
of ovnkube-master.

We make it configurable and assign it a verbosity of 5. Default on
OpenShift is 4, however this log is redundant with the
"transacting operations" that is already logged with level 4.

Signed-off-by: François Rigault <frigo@amadeus.com>