E2E (exgw): create ns for the serving gw pods instead of using default
In environments where PodSecurity is enforced we can't use the default ns unless
certain labels are set. Namespaces we create with the framework have these labels
which let us create pods on them without restrictions.
Since we're using disposable namespaces that are deleted in AfterEach there is
no longer need to delete the gw pods.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>