Fix race in unit test between IP release and pod retry
On CI, releasing an IP address of a completed pod might take some extra time, while the code in our unit tests might already want to force a retry to add a different pod and expect it to be assigned the released IP address. This fix lets the retry logic run for a minute, while before it only ran once, which was not sufficient.

Signed-off-by: Riccardo Ravaioli <rravaiol@redhat.com>