deleteEgressIPAssignments must add SNATs only for relevant pods
deleteEgressIPAssignments was adding back the SNAT towards
nodeIP for all the egressIP managed pods irrespective of
whether the pod was managed by the egressIP being removed or
not.

This PR also changes addPerPodGRSNAT to only add back the SNAT
towards the nodeIP if the pod is on the same node as the
egressIP was on.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>
Co-authored-by: Patryk Diak <pdiak@redhat.com>