Remove podIPs cache from podAssignmentState
We were copying podIPs information from logicalPortCache
onto podAssignmentState cache. Reasoning was:
// To avoid a lot of duplicated
// work on ovnkube-master restarts when all egress IP handlers will most likely
// match and perform the setup for the same pod and status multiple times over.

So if the podKey was already in cache we were simply reusing it.
This causes race bugs around stateful sets. This PR removes the
use of egressIP pod cache and uses logicalPortCache always.
Drawback of this is irrespective of what's in the EIP cache, we
will always fetch the pod IP info from the logicalPortCache
so there might be a slight performace degradation (hypothesis, not
tested).

We are also adding a new util.GetPodIPs which can help us get
the podIPs given a pod kapi which is needed during teardown of
egressIPs since we no longer have a cache.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>