Remove podIPs cache from podAssignmentState
We were copying podIPs information from logicalPortCache
onto podAssignmentState cache. Reasoning was:
// To avoid a lot of duplicated
// work on ovnkube-master restarts when all egress IP handlers will most likely
// match and perform the setup for the same pod and status multiple times over.

So if the podKey was already in cache we were simply reusing it.
This causes race bugs around stateful sets. This PR removes the
use of egressIP pod cache and uses logicalPortCache always.
Drawback of this is irrespective of what's in the EIP cache, we
will always fetch the pod IP info from the logicalPortCache
so there might be a slight performace degradation (hypothesis, not
tested).

We are also adding a new util.GetPodIPs which can help us get
the podIPs given a pod kapi which is needed during teardown of
egressIPs since we no longer have a cache.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>
(cherry picked from commit df0f90b27302b93a63d4138610dc1a71d0af0357)

Conflicts in 4.11:
	go-controller/pkg/ovn/egressgw.go
because
https://github.com/ovn-org/ovn-kubernetes/commit/dbe0a9137698088b0516acdafba9e38a9853d699#diff-aae5aec1d5a36916dcc466832ab980fa0db49e1fb794a14b1416cdcf9df7b5c3
is missing downstream
	go-controller/pkg/util/pod_annotation.go
because
https://github.com/ovn-org/ovn-kubernetes/commit/6f6d134f04ac2a9ff9b44ccebcbb0bceb579c2c5#diff-3d3fe81a31268662d021e60eb37f4a1f7e96ddb7e61de7f49acbbd2ae82e81ca
is missing downstream