Don't use ACL names ever!
Initially when the commit to cleanup ACLs was done in
https://github.com/ovn-org/ovn-kubernetes/pull/3038 I had assumed
that ACL names weren't truncated to fit the 63 char length and
thought they were like other fields. Based on that assumption there
were two places in code where I used a predicate match based on
acl names which was plain stupidity. This PR fixes that:

1) When the arp && arp || nd bug was being cleaned up, I was
trying to be extra careful in matching on old exp: "arp" and
the acl name. I deleted the second part of that logic since
there can be cases where the entire suffix is missing like
when the namespace is 63 chars. So let us stick to matching
on acl.Match alone

2) When the deny default ACL duplicates per namespace were
being updated - again I was matching on ACL names to exclude
the arp policies. I replaced this with using acl.Match which
is more reliable and accurate.

In https://github.com/ovn-org/ovn-kubernetes/pull/3181 we realised
we had to trim the acls and we fixed that, but didn't fix the
above two issues. So let's fix that up and also I added a unit
test case with a long named namespace.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>