Disable packet MTU check when OVS HW Offload is enabled
This is a OVS dependent work around fix. If we detect that OVS HW Offload
is enabled, we are aware that the offloading of packet MTU check support
in OVS is not supported. This support will not be available until the offload
support for sFlow is added.

In order to have the most optimal performance, the packet MTU check
support should be disabled when OVS HW Offload is enabled on the node.

Signed-off-by: William Zhao <wizhao@redhat.com>