fixes to windows testing
general fixes to the windows node_linux.go testing to make the testing
more meaningful. The tests are not ideal because there is no interface
into fakeCmd to get the bundled ovs commands but we can compare the
cache to what we believe it should be.

I also removed tests that where no longer meaningful

test "removes stale node flows on initial sync" and "removes stale pod
flows on initial sync" are no longer needed because we sync to the
flowCache on every sync and that stale nodes/pods will be taken care
of then.

test "sets up local pod flows" is a duplicate of "sets up local linux
pod" so I removed the former

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>