egressip node update: set the node as reachable only when it is being added as an egressip node
there are three places where ovnkube marks an egress node as reachable:
 - node add event handler
 - node update even handler
 - periodic connectivity checker

In a scenario where the egressip node becomes reachable at the same time as the node object is updated
the node is set as reachable but because there was no label/readiness change it is never added as an egressip node.
The periodic check will then ignore the node because it considers it as already reachable.

To address the issue only set the node as reachable if it is actually getting added as an egressip node.

Signed-off-by: Patryk Diak <pdiak@redhat.com>