Handle expired entry while handling dns update
When nextQueryTime for dns entry is already expired and update goroutine is trying
to use negative duration to reset ticker. This causes process crash on the ovnk
master. This change avoids proces crash and ensures to trigger ticker in shortest
possible time (1 ms) so that update happens immediately on expired dns entry.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>