node: mock conntrack delete operations for service and EgressIP testcases
If real conntrack operations are performed against the CI host we may
get operation errors because we're not running a real node in the testcases
and we fake out all the iptables stuff too. We shouldn't mix real operations
in the CI host's netns with fake operations in the testscases because then we
cannot control the test parameters.

Second, don't start the node IP handler's netlink watch or sync host IPs
in testcases because it scrapes the CI host's IP addresses which service
deletion uses when clearing conntrack, which the testcases know nothing
about. Just stick with the fake node IP that the testcases already add.

This all fixes errors like:

[91m[1m• Failure [0.114 seconds][0m
Node Operations
[90m/go/src/github.com/openshift/ovn-kubernetes/go-controller/pkg/node/gateway_localnet_linux_test.go:161[0m
  on delete
  [90m/go/src/github.com/openshift/ovn-kubernetes/go-controller/pkg/node/gateway_localnet_linux_test.go:1211[0m
    [91m[1mdeletes iptables rules with ExternalIP [It][0m
    [90m/go/src/github.com/openshift/ovn-kubernetes/go-controller/pkg/node/gateway_localnet_linux_test.go:1212[0m

    [91mUnexpected error:
        <*errors.errorString | 0xc000de8cb0>: {
            s: "DeleteService failed for nodePortWatcher: failed to delete conntrack entry for service namespace1/service1: failed to delete conntrack entry for service namespace1/service1 with svcVIP 1.1.1.1, svcPort 8032, protocol TCP: operation not permitted",
        }
        DeleteService failed for nodePortWatcher: failed to delete conntrack entry for service namespace1/service1: failed to delete conntrack entry for service namespace1/service1 with svcVIP 1.1.1.1, svcPort 8032, protocol TCP: operation not permitted
    occurred[0m

    /go/src/github.com/openshift/ovn-kubernetes/go-controller/pkg/node/gateway_localnet_linux_test.go:1243
[90m------------------------------[0m
[0mNode Operations[0m [90mon delete[0m
  [1mdeletes iptables rules for NodePort[0m
  [37m/go/src/github.com/openshift/ovn-kubernetes/go-controller/pkg/node/gateway_localnet_linux_test.go:1295[0m

Signed-off-by: Dan Williams <dcbw@redhat.com>