CARRY: Fixes hybrid overlay pod add unit test
In upstream the ignoreLearn flag for flows installed is always true,
however ICNIv1 always resets the ignoreLearn flag to false after syncing
in order to make sure we learn the flow on the next sync after
installing our new base flow.

In upstream there is no real point of having ignoreLearn as we do not
install and learned action flows upstream. Will open a github issue
upstream to have this code removed.

Signed-off-by: Tim Rozet <trozet@redhat.com>