Reduce risk of overwriting default pod network annotation
If the pod informer cache is out of date, we may
replace an existing pods 'default' network annotation.

This was seen in the field where OVN-K master attempts
to set a pods default network annotation but it timed
out. However, it succeeded.
OVN-K later handled another event for this pod, looked
into the pod informers cache, saw no network set in
the annotation of the pod and generated a new default
network annotation and replaced the original network
annotation. Meanwhile, ovnkube node was waiting
for the annotation to be set and starting
configuring the network for the first default network
annotation that was configured.

This change attempts to get the latest pod information
from the kapi in order to reduce the risk of replacing
an existing network assigned to the pod.

Reported-at: https://issues.redhat.com/browse/OCPBUGS-3482

Signed-off-by: Martin Kennelly <mkennell@redhat.com>