sharing port_groups for multiple policies of the same local pod selector
Today, port group is created for per network policy, even if they share
the same local pod selector. This means for the same pod that is
selected, its logical port will be added/deleted for multiple port
groups.

This commit is to add support of sharing of port_groups for multiple
network policies if they have the same local pod selector.

Signed-off-by: Yun Zhou <yunz@nvidia.com>