Avoid duplicate address set transactions
When pod updates happen, handlers may be invoked that add pods IPs to
address sets. A prime example of this is network policy handlePeerPods.
Each time the pod is updated, the address will be added to the address
set with a mutate-insert operation. The cost of this txn when there are
many network policies that apply to this pod is high. For example, with
a pod that has 730 network policies, and 7 pod updates that would result
in 5110 transactions instead of only the 730 needed to add this pod in
the first place.

Signed-off-by: Tim Rozet <trozet@redhat.com>