Set gateway-chassis when LRP is being created, to avoid scaling problems.
Commit 09af6537 moved the LRP creation from ensureNodeLogicalNetwork to
syncNodeClusterRouterPort, so that when creating LRP the gateway-chassis
can be set, to pin LS to a chassis. However, Commit c58554adfc
accidently reverted the change when converting nbctl to libovsdb (it is
likely a rebase problem). It creates LRP in ensureNodeLogicalNetwork
without setting gateway-chassis, which end up with all existing
ovn-controllers change monitor condition to monitor the new LS datapath.
And later, when ovnkube-node annotates the chassis-id, ovnkube-master
calls syncNodeClusterRouterPort, which updates the LRP and sets the
gateway-chassis, so all the ovn-controllers (except the new node) change
monitor condition again to remove the LS.

Such monitor condition changes come from every existing node, so the SB
DB become extremely busy refiltering all the existing lflows/PBs for all
the monitor conditions - O(M x N) problem. Even if condtional-monitoring
is disabled, this would cause unnecessary churns for all the
ovn-controllers.

This patch fixes the problem by:
1. Remove the LRP creation from the ensureNodeLogicalNetwork.
2. Combining the LRP creation and gateway-chassis setting into the same
   transaction. For this purpose the related APIs in router.go are
   refactored/simplified.

Signed-off-by: Han Zhou <hzhou@ovn.org>