Update egress node assignability on every egress node update
setNodeEgressAssignable sets an egress node as assignable when it exists in the allocator cache, otherwise it does nothing.
Currently the egress node update handler only calls setNodeEgressAssignable when it detects that the nodeEgressLabel was added/removed.
In cloud environment a node is added to the allocator cache after it gets annotatated with cloud.network.openshift.io/egress-ipconfig by CNCC.

In a scenario where a node is labeled for egress IP assignment before it is annoted by CNCC, it never gets set as assignable(isEgressAssignable).
To address that call setNodeEgressAssignable on every node update.

Signed-off-by: Patryk Diak <pdiak@redhat.com>