[rfe] keep dns stale entries in address_set
For records containing multiple IPs, DNS server may answer a frequently
changing subset of IPs from a fixed, larger pool of IPs. This causes
flows to "flap" as IPs are moving in and out of address_set after each
DNS answer, causing flows to be continuously added and removed.

To avoid that, if an entry has been seen returned by the DNS server we
keep it in the address_set even though the DNS server does not answer it
anymore.  This looses up the EgressFirewall rule, as some previous IPs
will keep being matched for a little while.

This partially addresses https://issues.redhat.com/browse/RFE-1355