addMasqueradeRoute: fallback to gateway interface IPs
There might be circumstances where the node status internal IPs
are not current (in our case, a dual-stack conversion procedure in
which kubelet is not configured with the newly acquired IPv6 address
and restarted, which in itself is a problem that would need to be
fixed).

While node status IPs are preferred to avoid issues with IPs that the
user or other components might arbitrarily add to the gateway interface,
fallback to an interface IP if we are missing it from the node status.

Also fix DPU. Masquerading should not be needed in the DPU (when
newSharedGateway is called with mode NodeModeDPU), only in the
host (where initGatewayDPUHost is use instead for NodeModeDPUHost).
Ref: https://github.com/openshift/ovn-kubernetes/blob/master/docs/design/dpu_support.md

Signed-off-by: Jaime Caamaño Ruiz <jcaamano@redhat.com>