poll "egress firewall policy functionality" tests
There is a rare flake seen in the test "Should validate the egress
firewall policy functionality against remote hosts" that can be
reproduced locally, but the condition does not remain with the
test pod. The check that fails will pass after reproducing and
bailing the test without cleanup.

This adds some small polling so that if the 'nc' check fails
initially, it will get at least 2 more retries to pass, before
marking the test case as failed.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>