Ensure routes on ovn_cluster_router are not duplicated
There can be a case where multiple routes exist on ovn_cluster_router
matching on the same IP prefix. For example:

            10.130.0.0/23                100.64.0.4 src-ip ecmp
            10.130.0.0/23                100.64.0.6 src-ip ecmp
            10.130.0.0/23                100.64.0.9 src-ip ecmp

In this case stale routes exist for old join subnet nexthops matching on
the same host subnet. OVN allows for these extra routes to be added and
considers them ecmp.

This commit modifies our behavior to ensure only a single route exists.
If more than one route for a predicate are found, the additional routes
are removed and then the leftover route is updated with the correct
fields.

A route is considered unique when it has the same prefix and policy type
(dst-ip vs src-ip) on the same router. The only case in OVNK where we
will allow more than one identical route is for ECMP routes with
multiple external gateways. In this case it is intended for a single
prefix to have multiple nexthops. Note, egressgw code uses its own
method to generate the ops for creating the routes.

Signed-off-by: Tim Rozet <trozet@redhat.com>