Drop packets coming from pods headed externally that were not properly SNATed
Egress IP is often configured on a node different from the one hosting the affected pod.
Due to the fact that ovn-controllers on different nodes apply the changes independently,
there is a chance that the pod traffic will reach the egress node before it configures the SNAT flows.
Drop pod traffic that is not SNATed, excluding local pods(required for ICNI)

Signed-off-by: Patryk Diak <pdiak@redhat.com>