Fully remove dependency on default gateway for services
There remained a dependency for return service traffic from OVN->Host
where OVN would need to think it was routing towards the default gw.
Once this traffic made it into the shared gw bridge, it would be
hijacked and redirected back to the Host. This only applied to clusters
where a default gateway exists. In a cluster with no default gateway,
the default route for OVN pointed to a fake gw (169.254.169.4) with a
static MAC binding for this address.

However, in a cluster where there is a default gateway, but we are
unable to ARP and create a MAC binding, return service traffic will fail
to be routed. This commit changes the behavior so that:

1. Dummy GW MAC bindings are always created
2. Always add a route destined to the dummy masquerade subnet via dummy
   GW. Like "169.254.169.0/29 via 169.254.169.4"

This removes depending on an external default gateway for routing return
service traffic from OVN.

Signed-off-by: Tim Rozet <trozet@redhat.com>