Delete stale egress ip snat entries by node
There is an unknown scenario where stale pod snat entries are found
for egress ip which refers to an unassigned node. This change ensures
those stale entries are deleted in sync stale logic and also making
sure that egress ip reconcile does not break because of this entry,
otherwise it may lead to "unexpectedly found multiple results for
provided predicate" error in DeleteNATsWithPredicateOps.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>