loadbalancer: Disable generation of neighbor responder flows for VIPs on LRs
NB.Load_Balancer.options:neighbor_responder controls generation of
logical flows for routers in order to reply to ARP/ND request for a VIP.

Default value is 'reachable' that means that routers on which the
load balancer is applied should reply to ARP/ND requests only for VIPs
that are part of a router's subnet.

However, there seems to be no such VIPs in ovn-k setups.  And the
function that is looking for them is one of the heaviest functions
in ovn-northd that can take up to 30% of the northd processing.

Switching this option to 'none' should noticeably reduce load on northd
in LB-heavy scenarios.

The 'none' option is available since OVN 22.12, older OVN versions will
just ignore it using a default 'reachable' value.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>